Homework 1
Inserted doctype, header, and body. Labeled the tab to say BUS36110:
App Dev Homework 1 (a little bit personalized…). Also, at the bottom,
included proper closures and links as in the target. I struggled a
little bit removing the dash, and then it hit me that all I had to do
was delete &mdash.